requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
h5py==3.7.0
idna==3.4
numpy==1.23.4
pathlib==1.0.1
Pillow==9.2.0
requests==2.28.1
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
typing==3.7.4.3
typing_extensions==4.4.0
urllib3==1.26.12
